feature_extract
delete a directory,add a  file
